+++ database/migrations/2016_06_01_000004_create_report_joins_table.php
@@ -4,9 +4,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateReportJoinsTable extends Migration
+return new class extends Migration
 {
-
     /**
      * Run the migrations.
      *
@@ -35,5 +34,4 @@ class CreateReportJoinsTable extends Migration
     {
         Schema::dropIfExists('report_joins');
     }
-
-}
+};