+++ database/migrations/2016_06_01_000005_create_report_middleware_table.php
@@ -4,9 +4,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateReportMiddlewareTable extends Migration
+return new class extends Migration
 {
-
     /**
      * Run the migrations.
      *
@@ -31,5 +30,4 @@ class CreateReportMiddlewareTable extends Migration
     {
         Schema::dropIfExists('report_middleware');
     }
-
-}
+};